requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.36
apscheduler==3.10.4
pytz==2023.3
python-dateutil==2.8.2
